+++ packages/atomic/src/components/commerce/atomic-commerce-search-box/atomic-commerce-search-box.new.stories.tsx
@@ -2,25 +2,43 @@ import type {
   Decorator,
   Meta,
   StoryObj as Story,
-} from '@storybook/web-components';
+} from '@storybook/web-components-vite';
+import {getStorybookHelpers} from '@wc-toolkit/storybook-helpers';
 import {html} from 'lit';
 import {wrapInCommerceInterface} from '@/storybook-utils/commerce/commerce-interface-wrapper';
 import {parameters} from '@/storybook-utils/common/common-meta-parameters';
-import {renderComponent} from '@/storybook-utils/common/render-component';
 
-const {decorator, play} = wrapInCommerceInterface({skipFirstRequest: true});
+const {events, args, argTypes, template} = getStorybookHelpers(
+  'atomic-commerce-search-box',
+  {excludeCategories: ['methods']}
+);
+const {decorator, afterEach} = wrapInCommerceInterface({
+  skipFirstRequest: true,
+  includeCodeRoot: false,
+});
 
 const normalWidthDecorator: Decorator = (story) =>
-  html` <div style="min-width: 600px;">${story()}</div> `;
+  html` <div style="min-width: 600px;" id="code-root">${story()}</div> `;
 
 const meta: Meta = {
   component: 'atomic-commerce-search-box',
   title: 'Commerce/Search Box',
   id: 'atomic-commerce-search-box',
-  render: renderComponent,
+  render: (args) => template(args),
   decorators: [normalWidthDecorator, decorator],
-  parameters,
-  play,
+  parameters: {
+    ...parameters,
+    actions: {
+      handles: events,
+    },
+  },
+  args: {
+    ...args,
+    'minimum-query-length': '0',
+  },
+  argTypes,
+
+  afterEach,
 };
 
 export default meta;
@@ -30,7 +48,7 @@ export const Default: Story = {};
 export const RichSearchBox: Story = {
   name: 'With suggestions, recent queries and instant products',
   args: {
-    'slots-default': ` <atomic-commerce-search-box-recent-queries></atomic-commerce-search-box-recent-queries>
+    'default-slot': ` <atomic-commerce-search-box-recent-queries></atomic-commerce-search-box-recent-queries>
       <atomic-commerce-search-box-query-suggestions></atomic-commerce-search-box-query-suggestions>
       <atomic-commerce-search-box-instant-products
         image-size="small"
@@ -41,7 +59,7 @@ export const RichSearchBox: Story = {
 export const StandaloneSearchBox: Story = {
   name: 'As a standalone search box',
   args: {
-    'attributes-redirection-url':
-      './iframe.html?id=atomic-commerce-search-box--in-page&viewMode=story&args=enable-query-syntax:!true;suggestion-timeout:5000',
+    'redirection-url':
+      './iframe.html?id=atomic-commerce-interface--with-product-list',
   },
 };