+++ packages/atomic/src/components/commerce/atomic-product-price/atomic-product-price.new.stories.tsx
@@ -1,13 +1,13 @@
-import type {Meta, StoryObj as Story} from '@storybook/web-components';
+import type {Meta, StoryObj as Story} from '@storybook/web-components-vite';
+import {getStorybookHelpers} from '@wc-toolkit/storybook-helpers';
 import {wrapInCommerceInterface} from '@/storybook-utils/commerce/commerce-interface-wrapper';
 import {wrapInCommerceProductList} from '@/storybook-utils/commerce/commerce-product-list-wrapper';
 import {wrapInProductTemplate} from '@/storybook-utils/commerce/commerce-product-template-wrapper';
 import {parameters} from '@/storybook-utils/common/common-meta-parameters';
-import {renderComponent} from '@/storybook-utils/common/render-component';
 
 const {
   decorator: commerceInterfaceDecorator,
-  play: initializeCommerceInterface,
+  afterEach: initializeCommerceInterface,
 } = wrapInCommerceInterface({
   skipFirstRequest: false,
   type: 'product-listing',
@@ -21,11 +21,12 @@ const {
       currency: 'USD',
     },
   },
+  includeCodeRoot: false,
 });
 
 const {
   decorator: commerceInterfaceDecoratorEUR,
-  play: initializeCommerceInterfaceEUR,
+  afterEach: initializeCommerceInterfaceEUR,
 } = wrapInCommerceInterface({
   skipFirstRequest: false,
   type: 'product-listing',
@@ -39,17 +40,32 @@ const {
       currency: 'EUR',
     },
   },
+  includeCodeRoot: false,
 });
 
-const {decorator: commerceProductListDecorator} = wrapInCommerceProductList();
+const {decorator: commerceProductListDecorator} = wrapInCommerceProductList(
+  'list',
+  false
+);
 const {decorator: productTemplateDecorator} = wrapInProductTemplate();
+const {events, args, argTypes, template} = getStorybookHelpers(
+  'atomic-product-price',
+  {excludeCategories: ['methods']}
+);
 
 const meta: Meta = {
   component: 'atomic-product-price',
   title: 'Commerce/Product Price',
   id: 'atomic-product-price',
-  render: renderComponent,
-  parameters,
+  render: (args) => template(args),
+  parameters: {
+    ...parameters,
+    actions: {
+      handles: events,
+    },
+  },
+  args,
+  argTypes,
 };
 
 export default meta;
@@ -60,7 +76,7 @@ export const Default: Story = {
     commerceProductListDecorator,
     commerceInterfaceDecorator,
   ],
-  play: initializeCommerceInterface,
+  afterEach: initializeCommerceInterface,
 };
 
 export const WithEURCurrency: Story = {
@@ -70,5 +86,5 @@ export const WithEURCurrency: Story = {
     commerceProductListDecorator,
     commerceInterfaceDecoratorEUR,
   ],
-  play: initializeCommerceInterfaceEUR,
+  afterEach: initializeCommerceInterfaceEUR,
 };