+++ packages/atomic/src/components/recommendations/atomic-recs-list/atomic-recs-list.new.stories.tsx
@@ -1,19 +1,29 @@
-import type {Meta, StoryObj as Story} from '@storybook/web-components';
+import type {Meta, StoryObj as Story} from '@storybook/web-components-vite';
+import {getStorybookHelpers} from '@wc-toolkit/storybook-helpers';
 import {html} from 'lit';
 import {parameters} from '@/storybook-utils/common/common-meta-parameters';
-import {renderComponent} from '@/storybook-utils/common/render-component';
 import {wrapInRecommendationInterface} from '@/storybook-utils/search/recs-interface-wrapper';
 
-const {decorator, play} = wrapInRecommendationInterface();
+const {decorator, afterEach} = wrapInRecommendationInterface();
+const {events, args, argTypes, template} = getStorybookHelpers(
+  'atomic-recs-list',
+  {excludeCategories: ['methods']}
+);
 
 const meta: Meta = {
   component: 'atomic-recs-list',
   title: 'Recommendations/RecsList',
   id: 'atomic-recs-list',
-  render: renderComponent,
+  render: (args) => template(args),
   decorators: [decorator],
-  parameters,
-  play,
+  parameters: {
+    ...parameters,
+    actions: {
+      handles: events,
+    },
+  },
+  args,
+  argTypes,
 };
 
 export default meta;
@@ -29,21 +39,22 @@ export const Default: Story = {
         </style>
         ${story()}`,
   ],
+  afterEach,
 };
 
-const {play: playNoFirstQuery} = wrapInRecommendationInterface({
+const {afterEach: afterEachNoFirstQuery} = wrapInRecommendationInterface({
   skipFirstQuery: true,
 });
 
 export const RecsBeforeQuery: Story = {
   tags: ['test'],
-  play: playNoFirstQuery,
+  afterEach: afterEachNoFirstQuery,
 };
 
 export const RecsWithFullTemplate: Story = {
   tags: ['test'],
   args: {
-    'slots-default': `<atomic-recs-result-template>
+    'default-slot': `<atomic-recs-result-template>
             <template>
               <atomic-result-section-visual>
                 <span>Visual Section</span>
@@ -72,12 +83,13 @@ export const RecsWithFullTemplate: Story = {
             </template>
           </atomic-recs-result-template>`,
   },
+  afterEach,
 };
 
 export const RecsOpeningInNewTab: Story = {
   tags: ['test'],
   args: {
-    'slots-default': `<atomic-recs-result-template>
+    'default-slot': `<atomic-recs-result-template>
             <template slot="link">
               <atomic-result-link>
                 <a slot="attributes" target="_blank"></a>
@@ -90,10 +102,12 @@ export const RecsOpeningInNewTab: Story = {
             </template>
           </atomic-recs-result-template>`,
   },
+  afterEach,
 };
 
 export const RecsAsCarousel: Story = {
   args: {
-    'attributes-number-of-recommendations-per-page': 4,
+    'number-of-recommendations-per-page': 4,
   },
+  afterEach,
 };