+++ packages/atomic/storybook-utils/search/search-interface-wrapper.tsx
@@ -2,29 +2,29 @@ import {
   SearchEngineConfiguration,
   getSampleSearchEngineConfiguration,
 } from '@coveo/headless';
-import {within} from '@storybook/test';
-import {Decorator, StoryContext} from '@storybook/web-components';
-import {html} from 'lit';
+import { Decorator, StoryContext } from '@storybook/web-components-vite';
+import { html } from 'lit';
 import type * as _ from '../../src/components.js';
+import { spreadProps } from '@open-wc/lit-helpers';
 
 export const wrapInSearchInterface = (
-  config?: Partial<SearchEngineConfiguration>,
-  skipFirstSearch = false
+  config: Partial<SearchEngineConfiguration> = {},
+  skipFirstSearch = false,
+  includeCodeRoot = true
 ): {
   decorator: Decorator;
-  play: (context: StoryContext) => Promise<void>;
+  afterEach: (context: StoryContext) => Promise<void>;
 } => ({
   decorator: (story) => html`
-    <atomic-search-interface data-testid="root-interface">
+    <atomic-search-interface ${spreadProps(includeCodeRoot ? { id: "code-root" } : {})}>
       ${story()}
     </atomic-search-interface>
   `,
-  play: async ({canvasElement, step}) => {
+  afterEach: async ({ canvasElement, step }) => {
     await customElements.whenDefined('atomic-search-interface');
-    const canvas = within(canvasElement);
     const searchInterface =
-      await canvas.findByTestId<HTMLAtomicSearchInterfaceElement>(
-        'root-interface'
+      canvasElement.querySelector<HTMLAtomicSearchInterfaceElement>(
+        'atomic-search-interface'
       );
     await step('Render the Search Interface', async () => {
       await searchInterface!.initialize({
@@ -43,12 +43,10 @@ export const wrapInSearchInterface = (
 
 export const playExecuteFirstSearch: (
   context: StoryContext
-) => Promise<void> = async ({canvasElement, step}) => {
-  const canvas = within(canvasElement);
-
+) => Promise<void> = async ({ canvasElement, step }) => {
   const searchInterface =
-    await canvas.findByTestId<HTMLAtomicSearchInterfaceElement>(
-      'root-interface'
+    canvasElement.querySelector<HTMLAtomicSearchInterfaceElement>(
+      'atomic-search-interface'
     );
   await step('Execute the first search', async () => {
     await searchInterface!.executeFirstSearch();