+++ utils/ci/npm-publish-package.mjs
@@ -26,7 +26,7 @@ async function isPublished(name, version, tag = version) {
 }
 
 const isPrerelease = process.env.IS_PRERELEASE === 'true';
-const tagSuffix = process.env.TAG_SUFFIX || '';
+const tagSuffix = process.env.PR_NUMBER || '';
 /** @type {import('@npmcli/package-json').PackageJson} */
 const {name, version} = JSON.parse(
   readFileSync('package.json', {encoding: 'utf-8'})